Refactor: Split out PrimitiveNode into separate files
